booking_agent/agent/agent_config.py
from agents import AsyncOpenAI, OpenAIChatCompletionsModel , Agent , Runner  , set_tracing_disabled
from dotenv import load_dotenv
import os

load_dotenv()
set_tracing_disabled(disabled=True)
API_KEY = os.getenv("API_KEY")

provider = AsyncOpenAI(
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/",
    api_key=API_KEY
)

model = OpenAIChatCompletionsModel(
    model = "gemini-2.0-flash",
    openai_client = provider
)
