booking_agent/pyproject.toml
[project]
name = "roadmap-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "openai-agents>=0.1.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = ["backened"]
